create example node backend with http endpoints
I have created a placeholder backend using Node, ExpressJS, and
Typescript. It has example http endpoints demonstrating GET and POST.
It currently is entirely unrelated to the features that our eventual
application will use, and it may be replaced later with a different
back-end in a different language, but it will be helpful to test
communication between front and back-end while prototyping.