add push notification example
I have added an example implementation of a web push notification
subscription using angular's SWPush and the NodeJS web-push library.
Since we do not yet have any user data or the rest of the app
implemented, this demonstration only opens the app's homepage instead of
opening the "start chat" page.

To support this, I have added service worker/PWA support to the
front-end. Since service workers are not available in the `ng serve`
development server, I have added an npm script to build the app with
service workers enabled and serve it on a local http server.