add linting and rewrite readme
We now have automatic style checking using eslint; one set of rules is
shared by both the front-end and back-end code, but they are linted
separately. A .editorconfig file has been added to that supporting
editors will know to consistently use 2-space indentation and fix a few
other common issues like trailing whitespace.

I also moved the readme to the root of the project and added
instructions for how to build and run that backend and how to use the
linter.