allow sending non-object values through api
The request and response body types can now be any JSON-representable
value or plain text, rather than being restricted to only map objects
as it was during earlier prototyping. Strings can now be sent directly
rather than needing to be wrapped by objects; as before, non-string
values like objects or arrays will automatically be converted to and
from JSON by the Angular http client on the frontend and the expressJS
body parser middleware on the backend.