add backend support for authorization
adds backend endpoints for creating an account and logging in,
as well as a database table to support this. init.sql now creates a
special "admin" account, which will be able to send special instructions
to the backend for demos.

each user account is given a permanent UUID which is the primary key for
that user in the database, which will make it possible to easily change
an existing user's username without having to update multiple tables.

session tokens are provided to the client when the user logs in, and
will be sent as a header in requests made to the backend. the list of
active sessions will be remembered as long as the server is running but
is not persisted in the database, so it can be used to quickly (i.e.
synchronously) verify the identity and permissions of users. session
tokens are v4 UUIDs prefixed with 'session-'; the random UUID is
generated with the npm "uuid" package, which promises cryptographically
strong random values.

common code that will be shared by frontend and backend was extracted
to a separate "core" package, which now needs to be built separately.