fix lint configuration issues
1. the linter for boop-core was attempting to lint the typescript files
in `core/dist`, resulting in errors. this has been fixed so that eslint
will now ignore the core build output.

2. The linter for frontend failed because it could not parse
environment.prod.ts, since that replacement file was not part of tsc
compilation. this has been fixed so that all files are compiled,
including unused compilation files. This reintroduces a warning when
running `ng serve`, since it introduces an unused file into the build
output, but this warning can be safely ignored.