add token authentication and admin commands
the frontend's api service now sends the current session token in a
header, and the backend can use this to determine the identity of the
user.

also added an "admin commands" system, so actions on the backend
can be triggered from the browser console if the user is logged in
as an admin. note that the backend will reject commands that are sent
without an admin's session token.