improve database authentication issues
I added an optional argument for the backend that allows using a
different postgres username. The backend server will also stop
immediately if if fails to connect or authenticate, rather than not
showing an error until it receives a request. Should be useful on
non-windows platforms where the postgres installer does not create a
default user named "postgres".