keep header in place when scrolling
the header at the top of the page will now stay in place while
the rest of the app scrolls. also, the div containing the router-outlet
(i.e. the div containing the current page's component) will expand
to fill the available space if the current component is smaller,
making it easier for components to use layouts that expand to fill
empty space.