refactor admin into separate table
keeps track of which accounts are admins by a separate table of
admins' uuids rather than an "is_admin" column on each user. this scales
better since the number of admins should be trivially small compared to
the number of users, and is also makes queries for creating/editing
user accounts simpler.