implement sending boop reminderNotifications
the backend can now send notifications, and the frontend service worker
can respond to them by opening a page. the permissions of the person
viewing the notification and the identity of the person to start a chat
with are specified by a token that is sent in the payload of the
notification.