improve database efficiency for push subscriptions
we never want to get just one subscription at a time. we always want
all of the subscriptions for a specific user,so subscriptions are now
an array column of the user table rather than a separate table.

this makes adding new subscriptions slightly slower but makes retrieving
subscriptions both faster and simpler.