reduce number of queries to send a notification
improves efficiency of sending scheduled notifications. we no longer
need to separately query to get the vapid subscriptions of everyone we
send a notification to and to get the name/gender of everyone mentioned
in a notification message. instead, all of that is fetched in the same
query that selects who to send notifications to.