add icons and links for popular chat platforms
for the most popular chat platforms (i.e. the ones that have dedicated
options in the edit contact dropdown), we now show their icons
in the edit contacts page and the start chat page.

for some chat platforms, the start chat page can now link directly to
that platform, or even to the specific user's profile if the platform
supports a consistent url format for this.