fix 404s when navigating directly to pages
servers like http-server (used to test pwa mode locally) and github
pages assume that every page of our website has a different html file,
and show a 404 if given a path that doesnt match. Ours is a single-page
web-app, so we want all pages to be redirected to index.js while
preserving the href so that angular knows what page to show. we do this
by creating a custom 404.html that saves the href and redirects to
index.html