move notification pair selection to sql function
selecting possible pairs of users to send notifications to is a very
long sql query. previously we were sending the entire query from the
backend to the postgres server (more than 3 kilobytes!) every time.
making this a function should therefore save a lot of unnecessary
traffic between node and postgres.