fix issues with date processing
fixes bugs related to date parsing, such as the settings page
outputing date strings in JS format rather than ISO format.

adds luxon as a dependency to make date parsing safer.

fixes another problem on the settings page where it was not possible
to submit info settings when gender was set to "prefer not to say"