prevent icon masking/clipping
our current set of icon images was not designed to be used with
icon masking, so the app icon is cut off poorly on the android
homescreen. marking the icon as not maskable means that it will be
surrounded by white square/circle/squircle shapes on android, which
should look okay for our icon.